Moved error detection to DownloadManager
